Fronted/src/requirements.txt
pydantic
streamlit
python-dotenv